Add dependency instalation to the build command
Now, when running build, all of the folders and bind mounted and the
build file copied to the chroot.

After that, the helper executable is called so it can install all of
the build dependencies of the package.

There is a bug at the moment with the toml parser used, which assumes
python 2 strings on the dumps function. While this change isn't
corrected upstream, it will be required to patch it manually for the
program to work.